refactor: promote chat thread signals to agent-chat.ts base layer
Move chatThreads$, currentChatThread$, reloadChatThreads$,
reloadCurrentChatThread$, and ChatThread type from zero-chat.ts
to agent-chat.ts. Add re-exports in zero-chat.ts for backward
compatibility. Remove unused sidebarSubagents$.
